faircare-fl/requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
matplotlib==3.8.3
tqdm==4.66.2
PyYAML==6.0.1
torch==2.1.2
scipy==1.11.4
tabulate==0.9.0
